feat(scripts): Improvements
- Added options for scripts

+++ startup-service.sh
@@ -1,5 +1,26 @@
 #!/bin/bash
-echo "Start up the Chatboat services..."
-rm whatsapp/tokens/whatsapp-session/Singleton*
+
+echo "Start up the Analytics messenger services..."
+source .env;
+
+WHATSAPP_SINGLETON_FILES="whatsapp/tokens/whatsapp-session/Singleton*"
+
+if [ -d "$WHATSAPP_SINGLETON_FILES" ]; then
+  rm whatsapp/tokens/whatsapp-session/Singleton*
+fi
+
+
 docker compose up -d
-docker compose logs -f whatsapp
+
+echo "Services available at http://localhost:$PROXY_PORT"
+# Get the first argument and assign it to a variable $SHOULD_FOLLOW_LOGS
+
+SHOULD_FOLLOW_LOGS=$1
+if [ "$SHOULD_FOLLOW_LOGS" = "follow" ]; then
+  echo "Following logs"
+#  docker compose logs -f whatsapp
+  else
+    echo "To follow logs, run the script with the argument follow"
+    echo "Example: ./startup-service.sh follow"
+fi
+
